Cambios Locos
le movi unas cosas ahi sabe ahi que le movi, unos cambios a atributos o cosas que se llevan arrastrando desde cosas pasadas, una pulida mas que nada (por ejemplo un aitributo mas a producto que sea habilitado y deshabilitado), un metodo a productoDAO para agregar producto junto con ingredientes y otras cosas ahi

+++ Proyectos_Restaurante_252116_252595/RestauranteDominio/src/main/java/DTOS/Ingredientes/NuevoIngredienteDTO.java
@@ -11,12 +11,12 @@ package DTOS.Ingredientes;
 public class NuevoIngredienteDTO {
     private String nombre;
     private String unidad_medida;
-    private int stock;  
+    private double stock;  
 
     public NuevoIngredienteDTO() {
     }
 
-    public NuevoIngredienteDTO(String nombre, String unidad_medida, int stock) {
+    public NuevoIngredienteDTO(String nombre, String unidad_medida, double stock) {
         this.nombre = nombre;
         this.unidad_medida = unidad_medida;
         this.stock = stock;
@@ -38,11 +38,11 @@ public class NuevoIngredienteDTO {
         this.unidad_medida = unidad_medida;
     }
 
-    public int getStock() {
+    public double getStock() {
         return stock;
     }
 
-    public void setStock(int stock) {
+    public void setStock(double stock) {
         this.stock = stock;
     }
